Final update
This project has been put in the icebox for ever. An new version will be created of this program with new commands and a completely new codebase. The only changes that have been made to this program are changes so that you can start the program without it crashing instantly

-With kind regards
ItsADanny